Change FuncAnimation so that it only plots background once.
The background is plotted once and only the lines are cleared on each frame
of the animation. This should speed up the initial plotting - on my system
this descreased the plot time from 34s to about 20s, which was comparable
to the animation with no background at all.

The saving of the animation to the mp4 file seems to take around 35s on my
system regardless of the method used, and the files generated seem
identical.

Also added an alternate animation cell that updates the lines instead of
clearing them, which may of may not be slightly faster.

Finally, also minor tidy-ups to the matplotlib code, including legend
location, and just specifying number of frames and FuncAnimator will
automatically create an iterator.